Fix bug merge

+++ app/src/main/java/ca/unb/mobiledev/studyhub/FirebaseService.kt
@@ -458,7 +458,7 @@ object FirebaseService {
         }
     }
 
-    fun updateDayStudyTime(name: String, timeAdd: Long){
+    fun updateDayStudyTime(name: String, timeAdd: Double){
         val year = getCurrentYear()
         val week = getCurrentWeek()
         val day = getCurrentDayOfWeek()
